+++ .github/CODEOWNERS
@@ -270,12 +270,12 @@
 /comp/checks/winregistry @DataDog/windows-agent
 /comp/core/agenttelemetry @DataDog/agent-metrics-logs
 /comp/core/autodiscovery @DataDog/container-platform
-/comp/core/datadogclient @DataDog/container-platform
 /comp/core/sysprobeconfig @DataDog/ebpf-platform
 /comp/core/tagger @DataDog/container-platform
 /comp/core/workloadmeta @DataDog/container-platform
 /comp/metadata/packagesigning @DataDog/agent-delivery
 /comp/trace/etwtracer @DataDog/windows-agent
+/comp/autoscaling/datadogclient @DataDog/container-platform
 /comp/etw @DataDog/windows-agent
 /comp/rdnsquerier @DataDog/network-device-monitoring
 /comp/serializer/compression @DataDog/agent-metrics-logs
