+++ comp/autoscaling/datadogclient/component.go
