+++ comp/autoscaling/datadogclient/component_mock.go
