+++ comp/autoscaling/datadogclient/datadogclientimpl/singleclient.go
