+++ pkg/util/kubernetes/autoscalers/datadogexternal_test.go
@@ -16,8 +16,8 @@ import (
 	"github.com/stretchr/testify/require"
 	"gopkg.in/zorkian/go-datadog-api.v2"
 
-	"github.com/DataDog/datadog-agent/comp/core/datadogclient"
-	"github.com/DataDog/datadog-agent/comp/core/datadogclient/datadogclientimpl"
+	"github.com/DataDog/datadog-agent/comp/autoscaling/datadogclient"
+	"github.com/DataDog/datadog-agent/comp/autoscaling/datadogclient/datadogclientimpl"
 	"github.com/DataDog/datadog-agent/pkg/config"
 	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 	"github.com/DataDog/datadog-agent/pkg/util/pointer"
